added implementation for stock aggregate entity

+++ DTOs/Polygon/PolygonTickerAggregate.cs
@@ -1,5 +1,5 @@
 ﻿namespace BarCLoudTaskBackEnd.DTOs.Polygon
 {
-    public record PolygonTickerAggregate(int v, decimal vw, decimal o, decimal c, decimal h, decimal l, decimal t, decimal n);
+    public record PolygonTickerAggregate(int v, decimal vw, decimal o, decimal c, decimal h, decimal l, decimal t, int n);
 
 }
